Compile in evdev code to vxworksfb
Make it compatible with eglfs. The behavior is the same: By default
mouse, keyboard and touch will all be initialized and discovered
automatically. The environment variable QT_QPA_FB_DISABLE_INPUT
can be used to used to disable the built-in input handlers.

This allows to easily fall back from eglfs to vxworksfb on devices
that are not rendering via OpenGL.

[ChangeLog][QtGui] The vxworksfb platform plugin's input device handling
is now compatible with eglfs. The evdev keyboard, mouse and touch code
is compiled in by default.

Change-Id: I250118fa3a4f55cb20a1099e53e2dd541e7d686c
Reviewed-by: Laszlo Agocs <laszlo.agocs@theqtcompany.com>