596d69e04c merge: extra change in qglobal.cpp
596d69e04c merge added two vxworks-specific includes to qglobal.cpp.
The first one (with check for "_WRS_KERNEL" macro) looks incorrect, as
it was previously removed in 2f6e0e2153.

Note:
This commit has been added as the result of comparison between the head
of our branch with vxworks changes rebuilt as linear history and merge
commit from the vxworks-5.9 branch. It contains changes introduced
during the merge operation. We included all these changes on our
(linearized) branch so that heads of both branches are identical after
each rebase (and we can clearly see all changes introduced by merges).

Change-Id: I69d9252b93cd28dc5ee3f5bd71315221b8cc3cfc