596d69e04c merge: side effect from 7f9f2f6503
596d69e04c merge commit reverted the change from 7f9f2f6503
(probably as a result of mistake when resolving merge conflict).
It should be considered whether this change is to be removed.

Note:
This commit has been added as the result of comparison between the head
of our branch with vxworks changes rebuilt as linear history and merge
commit from the vxworks-5.9 branch. It contains changes introduced
during the merge operation. We included all these changes on our
(linearized) branch so that heads of both branches are identical after
each rebase (and we can clearly see all changes introduced by merges).

Change-Id: Ia79b59df10d9375d09dd4e33aa53daf6006a198f