a95917e034 merge: extra change in qnetworkinterface_unix.cpp
a95917e034 merge commit removed code defining QT_NO_IPV6IFNAME on
vxworks.
It looks like a bug during merge conflict resolution (there was a
conflict), as the code in question was added in 5c981b0b3e together
with #ifndef in the same file that used the macro, but the latter
was NOT removed here.
It should be considered whether this change is to be reverted.

Note:
This commit has been added as the result of comparison between the head
of our branch with vxworks changes rebuilt as linear history and merge
commit from the vxworks-5.9 branch. It contains changes introduced
during the merge operation. We included all these changes on our
(linearized) branch so that heads of both branches are identical after
each rebase (and we can clearly see all changes introduced by merges).

Change-Id: If52d7b19b61ba1c65d59a52221c310a5ecc0c8b2