Use math functions from qmath.h in examples
Math functions like sin, cos, sqrt, etc. are not available in the
default namespace on VxWorks, which was assumed in some code in the
examples, so they fail to compile on vxWorks.

As the code in question uses qreal type for parameters and output
variables when calling the functions, use the corresponding functions
from qmath.h (qSin, qCos, etc.), solving VxWorks compilation issue as
a side effect.

Change-Id: I78b07f2205a189c3c96677f4f88069ebeb9713b7
Reviewed-by: Tuomas Heimonen <tuomas.heimonen@qt.io>