pcre2: disable prce2 JIT on VxWorks
Disable JIT compilation of regular expressions on VxWorks (via pcre2)
as it leads to crashes and hangups.

In older versions of Qt, JIT compilation was not performed by default.
Only after 10 uses of an expression it was JIT-compiled by the
optimizePattern() function (this behavior could be controlled via
OptimizeOnFirstUsageOption and DontAutomaticallyOptimizeOption
options).

With 06af9a1e38 ("QRegularExpression: refactor pattern optimization")
this behavior was changed. Now JIT compilation is always performed
immediately after compiling a pattern, unless: a) QT_ENABLE_REGEXP_JIT
variable says otherwise or b) Qt is compiled in debug mode (QT_DEBUG).

After that change it turned out that even very simple applications
crash or hang on VxWorks because of prce2 JIT not working. So the buggy
JIT was exposed to virtually all applications, not only to heavy regexp
users.

Note, that prce2 JIT compilation was crashing also on vxworks-5.9, but
it wasn't immediately visible, as the crash occurred only after 10 uses
of single regular expression.

Disabling prce2 JIT on VxWorks fixes the issue.

Change-Id: Iee5b043a074419d019f288d7939b83909c7088b2
Reviewed-by: Tuomas Heimonen <tuomas.heimonen@qt.io>