VxWorks mouse: fix sending events
VxWorks specific implementation of QEvdevMouseHandler::readMouseData
(in qevdevmousehandler_vxworks.cpp) has serious bug: the code is
designed such that different event types are handled inside the
"switch", updating xxxChanged flags accordingly, and then the code
after the switch calls sendMouseEvent() to actually emit an event.
The problem is that the second part was put outside Q_FOREVER instead
of inside and therefore it's DEAD code.

The only call to sendMouseEvent() that is actually used is located
inside "if (n < sizeof(EV_DEV_EVENT)) {", which "accidentally" worked
in the past, as read() returns 0 on vxworks when there're no more
events available.

Fix this by moving the code block at the end of the function into
Q_FOREVER block.

Change-Id: Ib744ddfb2185d2f09fc5983efbfce32a5ecb0ad0
Reviewed-by: Tuomas Heimonen <tuomas.heimonen@qt.io>