Workaround to avoid a bug from VxWorks
This is a workaround to avoid a bug from VxWorks.

Without the patch, it will cause the error,
"cannot initialize a parameter of type 'wchar_t *' with an rvalue
of type 'const wchar_t *'"

Since the std::wstring::data overload is not meant to be "const",
this is a compiler bug from VxWorks.

Change-Id: I51f711e1e574034af70676ca93dd710767413d07
Reviewed-by: Hyunkook Khang <hyunkook.khang@qt.io>
Reviewed-by: Pasi Petäjäjärvi <pasi.petajajarvi@qt.io>