Fix call to drawImage that causes a blank screen in VxWorks for x86
The call to drawImage in QPainter incorrectly passes the private member
mFbScreenImage from QVxWorksFbScreen, instead of mScreenImage from the
base class QFbScreen.

The code was originally correct. But, it was broken when QFbScreen was
refactored to reduce heap allocation by making mScreenImage a normal
member variable (daf5d3cf13). This would have caused
qvxworksfbscreen.cpp to fail to compile, as it now dereferenced a member
variable that was no longer a pointer.

This compilation failure was later "fixed" incorrectly by using the
similarly named mFbScreenImage instead (3883456ea3), which happened to
have the expected pointer to QImage type, rather than removing the
unwanted dereference of mScreenImage.

This commit reverts the code to the original behavior, while also
removing the dereference to accommodate the refactor of QFbScreen. This
fixes unwanted behavior in VxWorks for x86 that manifests as a
continuously blank screen when using the VxWorks frame buffer QPA.

Change-Id: I5d9f61f7d6be961795383e70cd656c19b179ae93
Reviewed-by: Pasi Petäjäjärvi <pasi.petajajarvi@qt.io>