Build integrated graphics support for x86 VxWorks
Hardware support for Intel integrated graphics was introduced to the
VxWorks build by changing the default platform plugin to EGLFS. This
required a reordering of the EGL shared libraries to address the
linker errors that were uncovered. It also brings in OpenGL-related Qt
source that is incompatible with MESA name mangling; So the
USE_MGL_NAMESPACE preprocessor macro has been removed to match a
corresponding change in the VxWorks kernel.

Change-Id: I92df27f1c8174772c1fd734d75518171ac7d73c3
Reviewed-by: Pasi Petäjäjärvi <pasi.petajajarvi@qt.io>