Address redefined getpagesize() in VxWorks headers
The VxWorks system headers define getpagesize() as a function in the
unistd.h header file. The same headers also define getpagesize() as a
macro in the xf86drm.h header file. Both definitions replace
getpagesize() with "(int)sysconf(_SC_PAGESIZE)". However, including
xf86drm.h before unistd.h results in a build failure because the
preprocessor will attempt to replace getpagesize(void) with its macro
definition. This cannot work, because "void" is interpreted to be an
argument by the function-like macro, which expects no argument.

To ensure that this situation does not occur, we check for existence
of the getpagesize macro before including unistd.h in qplatformdefs.h,
where it's expected to be found.

Change-Id: I41f02f63beb6d4f9a8c66a97bdf67088e8b8ad10
Reviewed-by: Pasi Petäjäjärvi <pasi.petajajarvi@qt.io>