Clear VxWorks MESA display artifacts on startup
Display artifacts are sometimes seen at startup in VxWorks applications
that make use of the MESA OpenGL libraries. This issue has been
demonstrated in the Wind River MESA OpenGL demos, which use the same
technique incorporated here to mask the problem. Cyling through the
complete set of framebuffers and clearing each during the initialization
process creates a clean slate that prevents the artifacts from being
noticable to the end user.

Change-Id: I00f89d33c69055d90fddadcdd031cdab9ce1089b
Reviewed-by: Pasi Petäjäjärvi <pasi.petajajarvi@qt.io>