QThread::currentThreadId: fix build on x32 (ILP32) ABI
On this ABI, pointers are 32-bit, so Qt::HANDLE (void *) is a 32-bit
variable and the "movq" instruction is inappropriate.

There's a GCC extended inline assembler modifier for the instruction
size suffix (%z0) but Clang seems not to understand it. Instead, I just
removed the suffix: we can do that because this is a memory load
instruction, which implies the destination is a general purpose register
(also required by the "=r" constraint) and therefore the assembler can
determine the size of the memory load from the name of the selected
register.

Note: I did not verify this compiles on x32 at all, much less that it
loads the right thing from memory.

On picking to 6.5, a conflict arose from a define-check on DARWIN that
has turned into a check on DARWIN64 since. As the code being modified
was the same in both cases, and the picked change did not meddle with
the #if-ery, I've just applied the same code-change within the
changd #if-ery.

Fixes: QTBUG-122674
Change-Id: I01ec3c774d9943adb903fffd17b6513d146e89ce
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit 08349ef0fe3902504987d12ebe0ed9674ed1e486)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit f5738ab46345adf17332d7cfd7aa238d27e09ed0)
(cherry picked from commit 71f10d4f4a50c40dba56503267f3991999fcda9f)
Reviewed-by: Marc Mutz <marc.mutz@qt.io>