Android: Restore previous foreign window behavior
A foreign window won't have a surface created for it, but we still want
the geometry to be updated for the wrapped native view. Same goes for
the visibility (fixed in Qt > 6.8 by 6faf18aceb21a4). Changes to
the foreign window and window embedding implementation in 6.7
introduced regressions in QtWebView, as QtWebView is one of the few
places that used the old window embedding code.
Most of these issue seems to been solved in 6.8, but those aren't
directly applicable to 6.7, so this is a 6.7 fix only.

Fixes: QTBUG-127530
Change-Id: I7aa0afd5566b48949e1270d58092b4d006e54ba2
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@qt.io>
(cherry picked from commit 050679afad118317ba685e84bb2a7c4908df91e2)