QStorageInfo/Unix: reset the valid state when doStat()ing again
If the user calls setPath() or refresh(), we are going to stat the
storage again. But neither retrieveVolumeInfo() nor initRootPath() unset
the flag if the storage became invalid.

This can't be tested automatically.

On picking to 6.5, only the change to qstorageinfo_unix.cpp is needed,
as the linux equivalent of the relevant part was just a copy (they
were split to disentangle other things). Also, remove 6.2 from Pick-to
as it's dead. Keep 5.15, although it'll probably never get taken in
there, as it's post-LTS, so it is available if sought.

Fixes: QTBUG-129689
Pick-to: 5.15
Change-Id: I8d4c5b4ed563e89f5723fffd2f550b169bb20884
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>
(cherry picked from commit e55d7fa1e09efc86e797efdf00656891f640d81d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 20bf08f01215f16894d2937a23edf725d7575d84)
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>