Add public QWindow API for safe area margins
The margins are exposed as a simple QMargins, as this is covers the
primary use-cases sufficiently at this point.

A Q_PROPERTY is not added, as this would conflict with the explicit
Qt Quick API that will also be added.

Task-number: QTBUG-125374
Change-Id: I504c3000473de0b09272aa5a3e58ebf9a41aa942
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
(cherry picked from commit 7707ab210fa0f1a478639e2bd05669697bb828e8)