QLocale: fix unintended copy [1/2]: rawStandaloneMonthName()
Coverity complained that 'name' was copied when it could have been
moved.

This is correct, but not trivial to understand: If, in a ternary
expression, one of the legs is an rvalue and the other is an lvalue,
then the ternary expression is also an rvalue. So the return statement
invokes the QString move constructor.

But, of course, the lvalue isn't being moved from in this case.
Instead, lvalue-to-rvalue conversion materializes a temporary QString
as a copy of lvalue, and _that_ is moved instead. This is what happens
with 'name' and why Coverity complains about a copy.

Note that this is purely due to the use of the ternary: If we write a
normal if, then the lvalue leg returns a local object and we get a
move from it for free (NRVO isn't possible, because it's not
statically known which of the two possible objects will be returned.

To fix, rewrite the code using normal a normal if statement.

Amends aa8393c94fea01a4806b204fd3aa343a4e90666b.

Coverity-Id: 425487
Pick-to: 6.5
Change-Id: Ia85346afc50d8a2a1abf8a5f079280c787cd964a
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit cd96362492375c50a9d0614b829c51eb6597d713)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 8624d36cc0b905baf03021bf484f21a9aa48cef0)
(cherry picked from commit f885f72b572ca33b7abece69655a74d99a6fc3fb)