util/unicode: port appendToSpecialCaseMap() from QList to QSpan
This makes the function independent of the actual input container
being passed, allowing us to port from QList to QVarLengthArray
step-by-step.

In readUnicodeData(), this allows to pass the single-element case
using initializer_list instead of a temporary QList.

Amends the start of the public history (but, to be fair, QSpan wasn't
available then).

As with readLineInto() in a previous patch, QSpan is not available in
Qt 6.5, but that's not a problem because this tool doesn't need to
compile with the Qt version it is generating code for, so we can use
Qt-latest.

Pick-to: 6.5
Change-Id: I10039af9d5b82a3d23fec451bf051a868db4c343
Reviewed-by: Ahmad Samir <a.samirh78@gmail.com>
(cherry picked from commit 98878bd87354be790f76713c467d002221fccf83)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 05ced03ef5fceac0a88c80ace89a9b5e073e8813)
(cherry picked from commit 2da05babe8f9ab7682fa1c89e170622d3c420343)