Doc: Fix ambiguity about QPainter::setOpacity's behavior
The documentation doesn't define what a "drawing operation" is. This
allowed two different interpretations how a shape with an outline and a
fill color will be drawn when 0 < opacity < 1:

- QPainter should draw the fill with the given opacity, then draw the
outline on top of it with the same opacity, i.e. the fill's edge can be
seen through the outline. This is what QPainter does and I recently
added a test for this behavior.

- QPainter should draw an opaque fill with an opaque outline to some
buffer. It should then apply the given opacity to the result of that and
place it in front of the background, i.e. the background can be seen
through fill and outline but the fill's edge is hidden by the outline.

The added sentence makes clear the opacity will be applied to fill and
outline individually.

Change-Id: I1a81816de4fa7278ffdd881bf5fa984a9d638f89
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit a8c3ebbdbc64db8231d50e2a5abbbb756f150298)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 8c3de65bd442a9cc20254db4aeffece5a26a9af3)
(cherry picked from commit f40934afb6b17a875093b1f9fc1a9d78cdfdea67)