Mark QStringList as security-significant (=default)
While QStringList contains a lot of specialized algorithms on not only
the list itself, but also the QString members, a reasonable user will
expect them to be convenience calling the resp. functions for each
member.

As such, the security implications inherent to QString manipulation
are considered firewalled off by QString (which is one of the reasons
to mark QString as security-critical: so that higher-level code can
assume it's "doing it's job"), and what remains is a QList, which is
(or ought soon to be) marked as significant-only.

All of the above: ditto QByteArrayList.

Task-number: QTBUG-135195
Change-Id: I6c791ba59d1727d413f553d6498804431451de4c
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
(cherry picked from commit 3ba8e2fbc483f8f5564bb30ee7437a9150e193b5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 0c50f59bf966b3acbb1d1763be3461b6601fb409)