[docs] Q_CONSTINIT: mention it attaches to definition
The macro (like the attributes and the C++20 keyword that back it)
belongs on the definition, not on the declaration, of the variable,
because it's a static assertion on properties of its initialization.

Some compilers apparently accept it on a non-defining declaration (or
we only checked in C++17 builds with old compilers), but we have
several variables in QtCore that are marked as described in this
documentation addendum, and they've not caused problems in the CI all
these years, so document as practiced, and ignore other combinations
that theoretically may work, too (on some compilers, at least).

Amends 80b6bcc385863a38d499685af9b8e857e46cbdba.

Pick-to: 6.5
Change-Id: I825127ff9ade483584f399fda78b7c7082f68ed8
Reviewed-by: Ivan Solovev <ivan.solovev@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>
(cherry picked from commit 6d444dfd900e1742a961fabffa6517f912eb01c4)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 9786c4a43c80d33328df3b6722f526b7031109f4)