QArrayData: fix exception safety in assign()
As mentioned at the top of assign(), the function ought to provide the
basic exception guarantee.

We were checking whether the construction of T from the result of the
projection is noexcept, but, as Thiago pointed out in a related
change, we ignored that the projection invocation can fail, too, in
which case we'd leave a corrupt container around (size would not match
the number of active objects in the list, holes in the middle if we
started to use the prepend buffer, but failed in the middle).

To fix, also check the projection for noexcept and fall back to the
existing non-noexcept-construction paths in that case.

It's pretty harmless, since the functionality isn't exposed in public
API (if you consider QArrayData* non-public, that is), and the only
user of projections was QString::assign(), and the projection it uses
cannot throw (cf. f2ea9d8dc8dc8d52490722f78ea46cd374d8e649).

Amends 782ccc6de5950ff1f6d3eeaaeacc7af4bc97a84f for the first hunk and
e07710007b4cf9c0665ac9d5c2b6c7ef588aae0a for the second.

Because of a massive code-move and -rewrite that has happened in dev,
I decided that it's simpler to fix this in 6.10 and forward-port to
dev, as it's one less conflict to resolve.

Fixes: QTBUG-141899
Change-Id: I026f703e5245422dce2951fd733178b5a0a4eefe
Reviewed-by: Dennis Oberst <dennis.oberst@qt.io>
(cherry picked from commit 73fa4cf2004690ae71ed9c41a93aa071fda98ecc)
Reviewed-by: Marc Mutz <marc.mutz@qt.io>